cmake: make library discoverable via find_package()
Library can be discovered through find_package(CONFIG) mode. Library
produces Trace::trace IMPORTED target to be linked against. The target
is discoverable both in installed state, and when the project is
included as subdirectory.

Additionally, a few minor cleanups are made:
- no need to look for C compiler, there are no C files
- no need to print information about C compiler (no need to print
  information about C++ compiler either, since this is done by CMake
  automatically, but since it exists already, it is probably for a
  reason)
- library version is specified (required for target versioning)
- enable_testing() is wrapped inside BUILD_TESTING, which is a standard
  CMake variable, and set to ON by default to preserve existing behavior

After installation (tested) or inclusion as subdirectory (not tested),
expected usage is:

    find_package(Trace CONFIG REQUIRED)
    target_link_libraries(foobar PRIVATE Trace::trace)

Library can be included via:

    #include <trace/trace.hpp>